chore: move GetResources to resources dir
