+++ ui/src/authorizations/containers/TokensIndex.tsx
@@ -7,7 +7,7 @@ import {Page} from '@influxdata/clockface'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 import LoadDataTabbedPage from 'src/settings/components/LoadDataTabbedPage'
 import LoadDataHeader from 'src/settings/components/LoadDataHeader'
-import GetResources from 'src/shared/components/GetResources'
+import GetResources from 'src/resources/components/GetResources'
 import TokensTab from 'src/authorizations/components/TokensTab'
 
 // Utils
