+++ ui/src/resources/components/GetResources.tsx
