+++ ui/src/telegrafs/components/TelegrafInstructionsOverlay.tsx
@@ -8,7 +8,7 @@ import {withRouter, WithRouterProps} from 'react-router'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 import {Overlay} from '@influxdata/clockface'
 import TelegrafInstructions from 'src/dataLoaders/components/verifyStep/TelegrafInstructions'
-import GetResources from 'src/shared/components/GetResources'
+import GetResources from 'src/resources/components/GetResources'
 
 // Constants
 import {TOKEN_LABEL} from 'src/labels/constants'
