+++ ui/src/variables/components/CreateVariableOverlay.tsx
@@ -5,7 +5,7 @@ import {withRouter, WithRouterProps} from 'react-router'
 // Components
 import {Overlay} from '@influxdata/clockface'
 import VariableFormContext from 'src/variables/components/VariableFormContext'
-import GetResources from 'src/shared/components/GetResources'
+import GetResources from 'src/resources/components/GetResources'
 
 // Types
 import {ResourceType} from 'src/types'
